refactor: home경로 수정
