+++ src/pages/index.tsx
@@ -1,23 +1,24 @@
-import React from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { useCallback } from "react";
-import * as userActions from "../store/modules/dropdown";
+import { S } from "./style";
 
 const Home = () => {
-  // const dispatch = useDispatch();
-  // const name = useSelector(({ user }: any) => user.name);
 
-  // const changeNickname = useCallback(() => {
-  //   dispatch(userActions.changeNickname("heeseong"));
-  // }, [dispatch]);
+  const onClick = async() => {
+    window.location.href = `${process.env.NEXT_PUBLIC_BASE_URL}oauth2/authorization/kakao`
+  }
 
   return (
-    <React.Fragment>
-      <h1>test</h1>
-      {/* <span>{name}</span> */}
-      {/* <button onClick={() => changeNickname()}>이름 변경</button> */}
-    </React.Fragment>
+    <S.Container>
+      <S.Wrapper>
+        <div>
+          알고리즘 실력 향상은 BAEKER!
+          <br />
+          알고리즘 공부를 위한 스터디 그룹을 만들고 가입하세요.
+          <br />
+          다양한 규칙 수행 보상을 통해 스터디 랭커에 도전해 보세요!
+        </div>
+        <S.KakaoButton onClick={onClick}/>
+      </S.Wrapper>
+    </S.Container>
   );
 };
-
 export default Home;
