+++ src/pages/style.ts
